Adding filter and converter for Salling stores

+++ backend/ApiApplication/ApiApplication/Database/Models/Product.cs
@@ -2,7 +2,7 @@
 
 public class Product
 {
-    public int EAN { get; set; }
+    public Int64 EAN { get; set; }
     public string Name { get; set; }
     public string Brand { get; set; }
     public double Unit { get; set; }
